requirements.txt
djangorestframework==3.14.0
Django==5.0.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
openapi==1.1.0